requirements.txt
llama-index-multi-modal-llms-openai
llama-index-vector-stores-qdrant
llama_index
ftfy
regex
tqdm
torch
torchvision
matplotlib
scikit-image
python-docx
PyMuPDF
Pillow
numpy
PyPDF2
pdf2image
pdfplumber
openpyxl
docx2txt
xlrd
streamlit
openai
llama-index-llms-openai
nltk
pypdf
python-dotenv
streamlit_pdf_viewer
docx2pdf
azure-storage-blob